model/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
transformers
torch
psycopg2-binary
sqlalchemy
pydantic
requests
beautifulsoup4
trafilatura
feedparser
pillow
numpy
scikit-learn
elasticsearch
python-dotenv
python-multipart
pytest
httpx
paddleocr